+++ R/bibentries.R
@@ -439,5 +439,15 @@ bibentries = c(
     pmid = "24027076",
     volume = "32",
     url = "https://onlinelibrary.wiley.com/doi/10.1002/sim.5958"
+  ),
+  bender_2018 = bibentry("article",
+    author  = "Bender, Andreas and Groll, Andreas and Scheipl, Fabian",
+    title   = "A generalized additive model approach to time-to-event analysis",
+    journal = "Statistical Modelling",
+    volume  = "18",
+    number  = "3-4",
+    pages   = "299--321",
+    url     = "https://doi.org/10.1177/1471082X17748083",
+    year    = "2018"
   )
 )